Extract `getOrPromptConfiguration` into `Command` extension
